fastapi_backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ML and image processing
onnxruntime>=1.16.0
# onnxruntime-gpu>=1.16.0  # Uncomment for GPU support
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0

# Data handling
pandas>=2.0.0

# Gradio UI
gradio>=4.0.0

# Web scraping (for nutrition)
selenium>=4.15.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI

# Development
python-dotenv>=1.0.0

# Supabase
supabase>=2.0.0
python-dotenv>=1.0.0